pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepdreaming"
version = "0.1.0"
dependencies = [
    "streamlit",
    "torch",
    "torchvision",
    "numpy",
    "matplotlib",
    "opencv-python-headless",
    "ipykernel",
    "hypothesis[pytest]",
    "pytest-cov",
    "black[jupyter]",
    "isort",
    "mypy",
    "ruff",
    "tox"
]
requires-python = ">= 3.10"

[tool.pytest.ini_options]
addopts = "--cov=deepdreaming"
testpaths = [
    "tests",
]

[tool.black]
line-length = 120
target-version = ['py312']

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 120

[tool.mypy]
disable_error_code = ["import-untyped"]
exclude = [
    ".venv/",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["deepdreaming", "app"]

[testenv]
deps = ["pytest", "hypothesis"]
commands = ["pytest"]
